+++ crates/air_r_formatter/tests/specs/r/while_statement.R
@@ -1,13 +1,9 @@
-while(a)a
-
 while(a){}
 
 while(a) {
   1 + 1
 }
 
-# TODO: Not entirely sure how this should be formatted.
-# It's not very common though.
 while({ complex }) {
   1 + 1
 }
@@ -16,30 +12,75 @@ while(super_long_function_name_is_true_man_this_is_a_really_really_long_function
   1 + 1
 }
 
-while(
-  # comment
-  a) {
-  1 + 1
+# ------------------------------------------------------------------------------
+# Autobracing
+
+while(a)a
+
+while(a) 
+  a
+
+# ------------------------------------------------------------------------------
+# Comments
+
+while # leads while
+(a) {
+  b
+}
+
+while
+# leads while
+(a) {
+  b
+}
+
+while(a # leads while
+) {
+  b
 }
 
 while(a
-# comment
+# leads while
 ) {
-  1 + 1
+  b
 }
 
-while(a # comment
-) {
-  1 + 1
+while( # leads while
+  a) {
+  b
 }
 
-while(a) # comment
+while(
+  # leads while
+  a) {
+  b
+}
+
+while(
+  a
+  # leads while
+) {
+  b
+}
+
+while(a) # leads b
+{
+  b
+}
+
+while(a)
+# leads b
+{
+  b
+}
+
+while(a) # dangles {}
 {}
 
-while(a) # comment
-1
+while(a) # leads b
+b
 
-while(a) # comment1
+while(a) # dangles {}
 {
-  # comment2
+  # dangles {} 2
 }